feat: delegate heavy bounding-box definition processing to service, create async comunication between download service and bounding-boxes extraction service
